start av potensiel split journalføring/distribuering
