+++ service/src/main/kotlin/no/nav/su/se/bakover/service/dokument/DokumentService.kt
@@ -8,7 +8,6 @@ import org.slf4j.Logger
 import java.util.UUID
 
 interface DokumentService {
-    fun journalførDokumenter()
     fun distribuer()
 }
