+++ service/src/main/kotlin/no/nav/su/se/bakover/service/dokument/JournalførDokumentService.kt
@@ -1,0 +1,5 @@
+package no.nav.su.se.bakover.service.dokument
+
+interface JournalførDokumentService {
+    fun journalfør()
+}
