+++ service/src/main/kotlin/no/nav/su/se/bakover/service/journalføring/JournalføringService.kt
@@ -1,0 +1,5 @@
+package no.nav.su.se.bakover.service.journalføring
+
+interface JournalføringService {
+    fun journalfør()
+}
