+++ service/src/main/kotlin/no/nav/su/se/bakover/service/skatt/JournalførSkattDokumentService.kt
@@ -1,0 +1,5 @@
+package no.nav.su.se.bakover.service.skatt
+
+interface JournalførSkattDokumentService {
+    fun journalfør()
+}
